+++ Internal.Roslyn44/WorkspaceAnalyzerOptionsInternals.cs
@@ -1,0 +1,19 @@
+using System.Composition;
+using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.Diagnostics;
+using MirrorSharp.Internal.Roslyn.Internals;
+
+namespace MirrorSharp.Internal.Roslyn43 {
+    [Shared]
+    [Export(typeof(IWorkspaceAnalyzerOptionsInternals))]
+    internal class WorkspaceAnalyzerOptionsInternals : IWorkspaceAnalyzerOptionsInternals {
+        public AnalyzerOptions New(AnalyzerOptions options, Project project) {
+            Argument.NotNull(nameof(options), options);
+            Argument.NotNull(nameof(project), project);
+            return new WorkspaceAnalyzerOptions(
+                options, project.Solution,
+                IdeAnalyzerOptions.GetDefault(project.Services)
+            );
+        }
+    }
+}